created controller/services/repo for other

+++ src/main/java/com/abc/onlinebanking/repository/TranscationRepository.java
@@ -2,6 +2,6 @@ package com.abc.onlinebanking.repository;
 import org.springframework.data.repository.CrudRepository;  
 import com.abc.onlinebanking.domain.*;
 
-public interface TranscationRepository extends CrudRepository<TranscationDetails, Integer>  {
+public interface TranscationRepository extends CrudRepository<TranscationDetails, Long>  {
 
 }
